+++ backend/src/model/dto/page_impl.rs
@@ -18,22 +18,4 @@ impl<T> Page<T> {
             total_pages: value.total_pages,
         }
     }
-
-    /// Used to convert the result of a pagination query to a page.
-    pub fn from_query_result(query_result: Vec<(T, i64)>, per_page: i32, page: i32) -> Self {
-        let total = query_result.get(0).map_or(0, |x| x.1);
-        let results = query_result.into_iter().map(|x| x.0).collect();
-        let extra_page = match total % i64::from(per_page) {
-            0 => 0,
-            _ => 1,
-        };
-        #[allow(clippy::cast_possible_truncation, clippy::integer_division)]
-        let total_pages = (total / i64::from(per_page) + extra_page) as i32;
-        Self {
-            results,
-            page,
-            per_page,
-            total_pages,
-        }
-    }
 }
